Final Andando
Tiene un solo error, es al eliminar el auto siendo atendido en caso de
que la cabina sola este libre, y alguna cabina sig este trabajando, para
probarlo descomentar el if al final del metodo nuevaLinea()

+++ src/java/Dominio/AutoCat1.java
@@ -46,7 +46,7 @@ public class AutoCat1 extends Auto {
 
     @Override
     public double tiempoAtencion(double RND) {
-        tiempoAtencion = 30;
+        tiempoAtencion = (double)30;
         return tiempoAtencion;
     }
